Resolves 6 conflicts and a duplicate-fuzz-generator collision:

- mdrp2350/src/threaded/emulator.rs: comment-only — took incoming wording.
- silicon_isr_diff_rp2040.rs: stale "ships ahead of silicon hookup" header
  cleaned up by origin's commit 21942b6 — took incoming.
- qemu_diff_m0plus.rs: kept HEAD's `m0plus_admits_wide` shared-helper
  structure (b9c6427's admit unification) AND incoming's `MASK_NZCV_ONLY`
  admit (8a98090's E.3 fix that prevents fuzz_m0plus_msr's APSR cases from
  being silently dropped). Kept both regression tests
  (filter_admits_mask_nzcv_only, emu_side_runs_msr_primask_without_panic).
- probe_diff_rp2040.rs: HEAD extracted the bin to `probe_diff_rp2040_lib`
  for `test_silicon_rp2040` reuse — took HEAD's thin-wrapper version, then
  applied incoming's `MASK_NZCV_ONLY` admit + regression test to the
  extracted library so the silicon and QEMU filters stay in sync.
- isr_scenarios_rp2040.rs: header status block — took incoming's concise
  V1+V2 description (matches the cleanup intent of 21942b6).
- mdrp2040/src/tests.rs: trivial whitespace.

Duplicate fuzz generator (lib.rs + thumb32_gen.rs): both branches
independently added an M0+ Thumb-32 fuzz generator with the same purpose.
Kept incoming's `generate_fuzz_m0plus_t32` (4-stage refinement: 6dec1c9
+ e6d2bbc + c10f5ee + 8a98090, including the QEMU-divergence narrowing in
c10f5ee), removed HEAD's `generate_fuzz_t32_m0plus_wide`. Renamed
`FUZZ-BOUNDARY:` to `FUZZ:BOUNDARY_BL` so the curated boundary cases
satisfy the `FUZZ:` prefix invariant. Updated the count test for the
single-generator total (15 boundary + 4·(count/4) randomised cases).

Tests: harness 557, mdrp2040 1039, mdrp2350 3123, all green.

+++ crates/mdrp2350/src/core/execute_thumb32.rs
@@ -337,7 +337,7 @@ impl CortexM33 {
                 let lsb = (((hw1 >> 12) & 0x7) << 2 | ((hw1 >> 6) & 0x3)) as u32;
                 let widthm1 = (hw1 & 0x1F) as u32;
                 let width = widthm1 + 1;
-                let val = (self.regs.r[rn] >> lsb) & ((1u32 << width) - 1);
+                let val = (self.regs.r[rn] >> lsb) & (((1u64 << width) - 1) as u32);
                 self.regs.r[rd] = sign_extend(val, width);
                 1 // M33 measured: 1 cycle
             }
@@ -346,7 +346,7 @@ impl CortexM33 {
                 let lsb = (((hw1 >> 12) & 0x7) << 2 | ((hw1 >> 6) & 0x3)) as u32;
                 let msb = (hw1 & 0x1F) as u32;
                 let width = msb - lsb + 1;
-                let mask = ((1u32 << width) - 1) << lsb;
+                let mask = (((1u64 << width) - 1) as u32) << lsb;
                 if rn == 15 {
                     // BFC: clear bits
                     self.regs.r[rd] &= !mask;
@@ -386,7 +386,7 @@ impl CortexM33 {
                 let lsb = (((hw1 >> 12) & 0x7) << 2 | ((hw1 >> 6) & 0x3)) as u32;
                 let widthm1 = (hw1 & 0x1F) as u32;
                 let width = widthm1 + 1;
-                self.regs.r[rd] = (self.regs.r[rn] >> lsb) & ((1u32 << width) - 1);
+                self.regs.r[rd] = (self.regs.r[rn] >> lsb) & (((1u64 << width) - 1) as u32);
                 1 // M33 measured: 1 cycle
             }
             // Undefined
